pyproject.toml
[project]
name = "paypal-azure-workshop"
version = "1.0.0"
description = "PayPal x Azure AI Workshop — Cracking the Ring"
readme = "README.md"
requires-python = ">=3.10"

# Core dependencies shared across all scenarios
dependencies = [
    # Azure OpenAI & AI Services
    "openai>=1.30.0",
    "azure-identity>=1.16.0",

    # Data Processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",

    # Environment Configuration
    "python-dotenv>=1.0.0",

    # HTTP & Utilities
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
# LLM modules (01-06): observability, telemetry, evaluation
llm = [
    "azure-ai-evaluation>=1.0.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "azure-monitor-opentelemetry>=1.4.0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b27",
    "rich>=13.7.0",
    "plotext>=5.2.8",
]

# Agents modules: MAF + Foundry Agent Service
agents = [
    "agent-framework>=1.0.0",
    "agent-framework-foundry>=1.0.0",
    "azure-ai-projects>=2.1.0",
    "rich>=13.7.0",
]

# Voice + CUA modules: Voice Live, audio I/O, CUA dependencies
voice = [
    "websockets>=12.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "Pillow>=11.1.0",
]

# Search + RAG modules: Content Understanding, AI Search
search = [
    "azure-ai-contentunderstanding>=1.0.0",
    "azure-search-documents>=11.6.0",
    "agent-framework>=1.0.0",
    "agent-framework-foundry>=1.0.0",
    "rich>=13.7.0",
]

# Everything — install all scenarios at once
all = [
    "paypal-azure-workshop[llm,agents,voice,search]",
]

# Development / testing
dev = [
    "pytest>=8.0.0",
    "paypal-azure-workshop[all]",
]

# Streamlit demo apps
demo = [
    "streamlit>=1.36.0",
    "paypal-azure-workshop[all]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils", "agents/utils", "Voice_CUA/shared"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
